The Application consists of Three Activities   1. Splash Screen Activity : This activity will judge that user has previously logged IN or not if not then it will redirect it to Login/Register Activity which is implemented at MainActivity    2. MainActivity : This actvity host 2 fragments on Tabs Layout through which user can login himself or register himself.  3. MovieListActivity : This Activity host 2 Fragment which have recyclerview in themselves which shows All Movies fetched from an API and Recommended Movies list based on User Like.
Tools : 
Volley : As mentioned in good To have section i have implemented the api calling using volley library. 
SQLite : Sqllite is used to store the authentication data.
 Glide : It is used to load the image from the URL in detched API

Note : Singleton class is used for the instance of the volley for api calling. DB helper is the class which consists the queries of the table creation and user authentication. LogInFragment/RegisterFragment consists code for the registration and authentication process AllmoviesFragment/RecommmendeMoviesFragment consists code for hosting of recyclerview of allmovies and recommended movies.

All the code is well commented and self explainatory
have a good day :)